progress in repository sql

+++ src/main/java/com/marlon/CleanArchTeste/Domain/Contracts/Repositories/User/UserCommandsRepository.java
@@ -9,6 +9,6 @@ public interface UserCommandsRepository extends CommandsRepository {
 
     void update(User user);
 
-    void deleteById(int id);
+    void deleteById(long id);
 
 }
